+++ src/xtopcom/xmbus/xevent_network.h
@@ -50,7 +50,7 @@ public:
 
     virtual int32_t get_class_type() const override {return xstatistic::enum_statistic_event_network;}
 private:
-    virtual int32_t get_object_size_real() const override {
+    virtual size_t get_object_size_real() const override {
         return sizeof(*this) + get_size(get_result_data()) + message.capacity()*sizeof(xbyte_t);
     }
 };
